rh-bkr: build package for beaker script in beaker task
Building NetworkManager during kickstart has the serious disadvantage
that we don't yet run the installed system. E.g. if the kernel has
issues there, it fails `make check` of the build process, although
the later run kernel is fine.

Also, a failure during kickstart does not allow to access the machine
later to inspect the failure reason (e.g. the testing logfiles).

Instead, move the building to a separate script and run it as a beaker
task.