platform: remove padding for IP address lifetimes
We used to pad the lifetime since the beginning (commit
f121995fad93eda886b2a34a8d79a45a5688b917).

However, there is not race involved, since our platform cache
is in sync with the messages from kernel (which didn't used to
be the case).

Also, when receiving a RA with a zero preferred time, we must
not extend the address lifetime by 5 seconds, but instead deprecate
the address immediately.

https://bugzilla.gnome.org/show_bug.cgi?id=763513